requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.24.4
pydantic>=2.0.0

# REST API
fastapi==0.115.6
uvicorn==0.32.1

# Load testing
locust==2.20.1